requirements.txt
pygifsicle==1.0.1
<<<<<<< HEAD
Pillow==8.3.0
=======
Pillow==8.1.1
>>>>>>> 252bbfe48973fcb390ea1cd4fe311dbda356be88
